requirements.txt
web: gunicorn "salty_app:create_app()"

[[source]]
url = "https://pypi.python.org/simple"
verify_ssl = true
name = "pypi"

[packages]
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
python-dotenv = "*"
requests = "*"
psycopg2-binary = "*"
pandas = "*"

[dev-packages]

[requires]
python_version = "3.8"
